Completed baseline CCA, changed utils.h functions to uint8_t. Need changes in GOL for uint8_t conversion

+++ include/kernels.hpp
@@ -2,6 +2,8 @@
 
 // TODO: add it to namespace
 #include <cstddef>
+#include <cstdint>
 
 void compute_next_gen(bool *currentGrid, bool *nextGrid, size_t N);
-void cyclic_compute_next_gen(int *currentGrid, int *nextGrid, int N);
+// void cyclic_compute_next_gen(int *currentGrid, int *nextGrid, int N);
+void cyclic_compute_next_gen(uint8_t *currentGrid, uint8_t *nextGrid, int N);
